% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{block-class}
\alias{block-class}
\title{@docType class
 @title Parent (virtual) class for all blocks (latent fields and observations)}
\description{
A block is the most basic component of a multi-variate spatio-temporal model. It can either consist of a latent field or an observations. Each block 
has a unique identifier (uid) and some cardinality n, the meaning of which depends on the specific class
}
\keyword{Block}
