% Generated by roxygen2 (4.0.0): do not edit by hand
\name{colour_graph_bfs}
\alias{colour_graph_bfs}
\title{Greedy graph colouring algorithm}
\usage{
colour_graph_bfs(adj.matrix, numcolours, startnode = 1)
}
\arguments{
\item{adj.matrix}{the graphs adjacency matrix (usually sparse)}

\item{numcolours}{the maximum number of colours to try out. The function gives an error if this is exceeded.}

\item{startnode}{the starting node for the BFS algorithm}
}
\value{
a data frame with the associated colour for each vertex (also \code{class})
}
\description{
This algorithm colours the nodes in sequential order, where the order is given through a breadth first search algorithm started on a
pre-specified start-node.
}
\author{
Andrew Zammit Mangion
}

