% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{ConstBasis-class}
\alias{ConstBasis-class}
\title{@docType class
 @title Constant functions basis}
\description{
ConstBasis inherits from the virtual class Basis. A set of constant basis functions is initialsied using the function \code{initConstbasis}.
}
\keyword{#'}
\keyword{@rdname}
\keyword{Basis}
\keyword{ConstBasisclass}
\keyword{functions}
