% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meshfns.R
\name{convert_INLA_mesh_MVST}
\alias{convert_INLA_mesh_MVST}
\title{Convert INLA mesh object MVST FEM class}
\usage{
convert_INLA_mesh_MVST(mesh)
}
\arguments{
\item{mesh}{The INLA mesh to convert (inla.mesh))}
}
\value{
MVST finite element mesh object
}
\description{
The MVST R package requires the FEM mesh components to be in a different object class 
to that of 'inla.mesh'. This function converts the 'inla.mesh' class objects to required 
MVST class, without export of the mesh to disk.
}
