% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{GMRF_basis-class}
\alias{GMRF_basis-class}
\title{@docType class
 @title GMRF_basis}
\description{
This class defines the amalgamation of a GMRF with a basis. It itself is a \code{process} block, i.e. it can be linked up with observations
using objects from the \code{link} class.
#' @rdname GMRF_basisclass
}
