% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{GMRF-class}
\alias{GMRF-class}
\title{@docType class
 @title GMRF}
\description{
A basic GMRF object which can also take an intrinsic value for future definitions (intrinsic GMRFs not implemented yet). This
class inherits from \code{process} which in turn is a block.
}
