% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{GRBFBasis-class}
\alias{GRBFBasis-class}
\title{@docType class
 @title Gaussian radial basis functions basis}
\description{
GRBFBasis inherits from the virtual class Basis. A GRBF basis functions is initialsied using the function \code{initGRBFbasis}.
}
\keyword{Basis}
\keyword{GRBF}
\keyword{functions,}
