% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{link_list-class}
\alias{link_list-class}
\title{@docType class
 @title List of links}
\description{
The list of links is simply a collection of links, each with its own incidence matrix. The link list is combined with a block list in order to 
compose a graph.
}
