\name{mvst-package}
\alias{mvst-package}
\alias{mvst}
\docType{package}
\title{
Bayesian inference for the multivariate Skew-t model.
}
\description{
Estimates the multivariate Skew-t (and nested) models under a Bayesian perspective using a Monte Carlo sampling method.
}
\details{
\tabular{ll}{
Package: \tab mvst\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2016-06-23\cr
License: \tab GPL (>= 2)\cr
}
}
\author{
Brunero Liseo, Antonio Parisi

Maintainer: Antonio Parisi <antonio.parisi@uniroma2.it>
}
\references{
Liseo, B. and Parisi, A. (2013) Bayesian inference for the multivariate skew-normal model: a population Monte Carlo approach, Computational Statistics and Data Analysis. doi:10.1016/j.csda.2013.02.007.
}
\keyword{ package }
