\docType{class}
\name{Obs-class}
\alias{Obs-class}
\title{observation block}
\description{
An observation block inherits from class \code{block},
however can only be used in certain ways in a graph. For
example it can be linked to an observation process block
but not to another observation block. To initialise use
\code{initObs}.
}
\details{

}

