% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{class}
\name{process-class}
\alias{process-class}
\title{@docType class
 @title Process block (virtual)}
\description{
A process block inherits from class \code{block}, however can only be used in certain ways in a graph. For example it can be linked to an observation
data block but not to another process block. It is the parent class of all latent processes.
}
\keyword{Block,}
\keyword{process}
