% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Meshfns.R
\name{write_INLA_mesh_MVST}
\alias{write_INLA_mesh_MVST}
\title{Export INLA Mesh}
\usage{
write_INLA_mesh_MVST(mesh, directory)
}
\arguments{
\item{mesh}{The INLA mesh to export, of class 'inla.mesh'}

\item{directory}{directory in which to save output files, must end with a '/'}
}
\value{
creates four csv files ‘p.csv, t.csv,
M.csv’ and ‘K.csv’ which are csv files containing the vertices,
triangulations, mass matrix and stiffness matrix details
respectively. The matrices are stored in the three-column format ‘i,j,k’.
}
\description{
Exports the 'inla.mesh' class object to CSV files that can be read by 
the 'Load_meshes' function of the MVST package. The MVST R package requires the FEM mesh 
components to be stored in a specific
csv file format in order to be properly imported and constructed.
}
