\name{mvst-package}
\alias{mvst-package}
\alias{mvst}
\docType{package}
\title{
\packageTitle{mvst}
}
\description{
\packageDescription{mvst}
}
\details{
\packageDESCRIPTION{mvst}
\packageIndices{mvst}
}
\author{
\packageAuthor{mvst}

Maintainer: \packageMaintainer{mvst}
}
\references{
Parisi A, Liseo B (2017). Objective Bayesian Analysis for the Multivariate Skew-t Model. Statistical Methods & Applications. ISSN 1613-981X. doi:10.1007/s10260-017-0404-0
Parisi, A., Liseo, B. (2018). Statistical Inference with Skew-t Distributions: the MVST R Package. Annali del Dipartimento di Metodi e Modelli per l'Economia il Territorio e la Finanza. ISSN 2385-0825.
Liseo, B., Parisi, A. (2013). Bayesian inference for the multivariate skew-normal model: a population Monte Carlo approach. Comput. Statist. Data Anal. <doi:10.1016/j.csda.2013.02.007>
}
\keyword{Skew-elliptical distributions}

