\name{mvst-package}
\alias{mvst-package}
\alias{mvst}
\docType{package}
\title{
\packageTitle{mvst}
}
\description{
\packageDescription{mvst}
}
\details{
\packageDESCRIPTION{mvst}
\packageIndices{mvst}
}
\author{
\packageAuthor{mvst}

Maintainer: \packageMaintainer{mvst}
}
\references{
Parisi A, Liseo B (2017). Objective Bayesian Analysis for the Multivariate Skew-t Model. Statistical Methods & Applications. ISSN 1613-981X. doi:10.1007/s10260-017-0404-0
}
\keyword{Skew-elliptical distributions}

